Also renamed 'diagnostic_ext' and 'session'.
