+++ server/src/diagnostic_handler.rs
